+++ src/main/java/gmail/davidsousalves/services/EntradaService.java
@@ -1,11 +1,12 @@
 package gmail.davidsousalves.services;
 
 import java.util.List;
-import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
+import gmail.davidsousalves.dto.EntradaDTO;
 import gmail.davidsousalves.model.Entrada;
 import gmail.davidsousalves.repositories.EntradaRepository;
 
@@ -19,15 +20,43 @@ public class EntradaService {
         return repository.findAll();
     }
 
-    public Optional<Entrada> findById(Long id) {
-        return repository.findById(id);
-    }
+	public EntradaDTO findById(Long id) {
+		Entrada entrada = repository.findById(id).orElseThrow(
+				() -> new IllegalArgumentException("Id nao existe"));
+		
+		return new EntradaDTO(entrada);	
+		
+	}
 
-    public Entrada save(Entrada entrada) {
-        return repository.save(entrada);
-    }
+	public EntradaDTO create(EntradaDTO entradaDto) {
+		Entrada entity = new Entrada();
+		copyDtoToEntity(entradaDto, entity);
+		entity = repository.save(entity);
+		return new EntradaDTO(entity);
+	}
+	
+	public EntradaDTO update(Long id, EntradaDTO entradaDto) {
+		Entrada entity = repository.getReferenceById(id);
+		copyDtoToEntity(entradaDto, entity);
+		entity = repository.save(entity);
+		return new EntradaDTO(entity);
+	}
 
-    public void deleteById(Long id) {
-    	repository.deleteById(id);
-    }
+	public void deleteById(Long id) {
+		if (!repository.existsById(id)) {
+    		throw new IllegalArgumentException("Recurso não encontrado");
+    	}
+    	try {
+    		repository.deleteById(id);    		
+    	}
+        catch (DataIntegrityViolationException e) {
+            throw new DataIntegrityViolationException("Falha de integridade referencial");
+        }	
+    	
+	}
+    
+    private void copyDtoToEntity(EntradaDTO dto, Entrada entity) {
+		entity.setDataEntrada(dto.dataEntrada());
+
+	}
 }
