+++ src/main/java/gmail/davidsousalves/services/SubGrupoProdutoService.java
@@ -1,11 +1,12 @@
 package gmail.davidsousalves.services;
 
 import java.util.List;
-import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
+import gmail.davidsousalves.dto.SubGrupoProdutoDTO;
 import gmail.davidsousalves.model.SubgrupoProduto;
 import gmail.davidsousalves.repositories.SubGrupoProdutoRepository;
 
@@ -19,17 +20,43 @@ public class SubGrupoProdutoService {
         return repository.findAll();
     }
 
-    public Optional<SubgrupoProduto> findById(Long id) {
-        return repository.findById(id);
-    }
+	public SubGrupoProdutoDTO findById(Long id) {
+		SubgrupoProduto subGrupo = repository.findById(id).orElseThrow(
+				() -> new IllegalArgumentException("Id nao existe"));
+		
+		return new SubGrupoProdutoDTO(subGrupo);	
+		
+	}
 
-    public SubgrupoProduto save(SubgrupoProduto subGrupoProduto) {
-        return repository.save(subGrupoProduto);
-    }
+	public SubGrupoProdutoDTO create(SubGrupoProdutoDTO subGrupoDto) {
+		SubgrupoProduto entity = new SubgrupoProduto();
+		copyDtoToEntity(subGrupoDto, entity);
+		entity = repository.save(entity);
+		return new SubGrupoProdutoDTO(entity);
+	}
+	
+	public SubGrupoProdutoDTO update(Long id ,SubGrupoProdutoDTO subGrupoDto) {
+		SubgrupoProduto entity = repository.getReferenceById(id);
+		copyDtoToEntity(subGrupoDto, entity);
+		entity = repository.save(entity);
+		return new SubGrupoProdutoDTO(entity);
+	}
 
-    public void deleteById(Long id) {
-    	repository.deleteById(id);
-    }
+	public void deleteById(Long id) {
+		if (!repository.existsById(id)) {
+    		throw new IllegalArgumentException("Recurso não encontrado");
+    	}
+    	try {
+    		repository.deleteById(id);    		
+    	}
+        catch (DataIntegrityViolationException e) {
+            throw new DataIntegrityViolationException("Falha de integridade referencial");
+        }	
+    	
+	}
 	
-	
+    private void copyDtoToEntity(SubGrupoProdutoDTO dto, SubgrupoProduto entity) {
+		entity.setDescricao(dto.descricao());
+		entity.setGrupoProduto(dto.grupoProduto());
+	}
 }
